requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.6
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
pycryptodome==3.14.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
